Moved CBlockFileInfo class from chain.h into its own header and updated .cpp and .h file dependencies.
Summary: Starts work on T380

Test Plan: `make check`

Reviewers: jasonbcox, deadalnix, schancel, O1 Bitcoin ABC, #bitcoin_abc

Reviewed By: jasonbcox, O1 Bitcoin ABC, #bitcoin_abc

Subscribers: teamcity

Differential Revision: https://reviews.bitcoinabc.org/D1654
