+++ src/txdb.h
@@ -6,6 +6,7 @@
 #ifndef BITCOIN_TXDB_H
 #define BITCOIN_TXDB_H
 
+#include "blockfileinfo.h"
 #include "chain.h"
 #include "coins.h"
 #include "dbwrapper.h"
